+++ packages/gamut/src/GridForm/GridFormSection/GridFormSection.tsx.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import { FieldError, useFormContext } from 'react-hook-form';
 
-import { GridFormInputGroup } from '../GridFormInputGroup';
 import { GridFormField } from '../types';
+import { GridFormQuestion } from './GridFormQuestion';
 
 export type GridFormSectionProps = {
   fields: GridFormField[];
@@ -15,29 +14,13 @@ export const GridFormSection: React.FC<GridFormSectionProps> = ({
   showRequired,
   pastFirstError,
 }) => {
-  const { register, errors, setValue } = useFormContext();
-
   return (
     <>
       {fields.map((field) => {
-        const errorMessage = (errors[field.name] as FieldError)?.message;
-        const isFirstError = !pastFirstError && errorMessage !== undefined;
-        pastFirstError = pastFirstError || isFirstError;
-        const requiredBoolean = !!(
-          field.type !== 'hidden' &&
-          field.type !== 'sweet-container' &&
-          field.validation?.required
-        );
-
         return (
-          <GridFormInputGroup
-            error={errorMessage as string}
-            isFirstError={isFirstError}
+          <GridFormQuestion
             field={field}
-            key={field.name}
-            register={register}
-            setValue={setValue}
-            required={requiredBoolean}
+            pastFirstError={pastFirstError}
             showRequired={showRequired}
           />
         );
