+++ packages/gamut/src/GridForm/GridFormSection/index.tsx
@@ -1,3 +1,4 @@
 export * from './GridFormSection.tsx';
 export * from './GridFormSectionTitle';
 export * from './GridFormSectionBreak';
+export * from './GridFormQuestion';
