+++ packages/styleguide/stories/Organisms/GridForm.stories.mdx
@@ -575,7 +575,7 @@ their own inputs. We can specify a 'custom' field type to with a [render prop](h
               </FormGroup>
             ),
             label: 'Gimme two more swags',
-            name: 'custom-input',
+            name: 'custom-input-group',
             validation: {
               required: true,
               pattern: {
